Fix compilation and warnings for QT example

+++ src/ui/qt/qt_ui.cpp
@@ -5,8 +5,8 @@
 UiWindowQt::UiWindowQt( const PenguinV_Image::Image & image )
     : UiWindow( image )
 {
-    const QImage imageQt( _image.data(), _image.width(), _image.height(), _image.rowSize(),
-                          (_image.colorCount() == 1u) ? QImage::Format_Grayscale8 : QImage::Format_RGB888 );
+    const QImage imageQt( _image.data(), static_cast<int>( _image.width() ), static_cast<int>( _image.height() ),
+                          static_cast<int>( _image.rowSize() ), (_image.colorCount() == 1u) ? QImage::Format_Grayscale8 : QImage::Format_RGB888 );
     _pixmap = QPixmap::fromImage( imageQt );
     _window.setPixmap( _pixmap );
 }
@@ -27,7 +27,7 @@ void UiWindowQt::drawPoint( const Point2d & point )
 {
     QPainter paint(&_pixmap);
     paint.setPen( QColor(20, 255, 20, 255) );
-    paint.drawPoint( point.x, point.y );
+    paint.drawPoint( static_cast<int>( point.x ), static_cast<int>( point.y ) );
     
     _window.setPixmap( _pixmap );
 	draw();
@@ -37,7 +37,7 @@ void UiWindowQt::drawLine( const Point2d & start, const Point2d & end )
 {
     QPainter paint(&_pixmap);
     paint.setPen( QColor(20, 255, 20, 255) );
-    paint.drawLine( start.x, start.y, end.x, end.y );
+    paint.drawLine( static_cast<int>( start.x ), static_cast<int>( start.y ), static_cast<int>( end.x ), static_cast<int>( end.y ) );
     
     _window.setPixmap( _pixmap );
 	draw();
